# Use Array.Clone() to copy arrays, and avoid copy where applicable