@drieseng 2018-10-14
If this approach is ok, I'll speed up the `Pattern(…)` overloads (that take an array) as well.