@JamesNK 2018-10-17
Where is the performance benefit coming from? Are there specific cases that are no longer copying, or is it from using Array.Clone?

I wonder why ToArray doesn't use Array.Clone internally if it is faster.