@drieseng 2018-10-17
Unrelated, but why doesn't the template for **MatcherAzureBenchmarkBase** combine multiple HTTP methods for a given endpoint in a single **RouteEndpoint** with a **HttpMethodMetadata** containing all HTTP methods for that endpoint? Wouldn't this be more representative?

Current we have this:

```cs
            ...
            Endpoints[138] = CreateEndpoint("/portalsettings/signin", "PUT");
            Endpoints[139] = CreateEndpoint("/portalsettings/signin", "PATCH");
            Endpoints[140] = CreateEndpoint("/portalsettings/signin", "GET");
            ...
```