@drieseng 2018-10-21
All PR feedback should have been addressed now.